+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The changelog for `Superwall`. Also see the [releases](https://github.com/superw
 - Tracks an `identity_alias` event whenever identify is called to alias Superwall's anonymous ID with a
 developer provided id.
 - Adds `setInterfaceStyle(interfaceStyle:)` which can be used to override the system interface style.
+- Adds `device.interfaceStyleMode` to the device template, which can be `automatic` or `manual` if 
+overriding the interface style.
 
 ### Fixes